Create armor.html
I realize this may not be necessary in the long run, but I wanted to add this in so that it could be used as a basic template. It also made sense to add the file to the repository itself, so that it's visible and stored online for future use.